feat: 에러 throw 부분 수정

+++ src/services/menus.service.js
@@ -1,5 +1,5 @@
 import bcrypt from 'bcrypt';
-import { InconsistencyError } from '../utils/err/err.js';
+import { InconsistencyError, NotFoundError, ValidationError } from '../utils/err/err.js';
 
 export class MenusService {
     constructor(menusRepository, usersRepository, restaurantRepository) {
@@ -18,7 +18,7 @@ export class MenusService {
         }
 
         if (!validation) {
-            throw { code: 400, message: '자신의 식당의 메뉴만 작성할 수 있습니다.' };
+            throw new ValidationError('자신의 식당의 메뉴만 작성할 수 있습니다.');
         }
 
         const menu = await this.menusRepository.createMenu(restaurantId, name, menuInfo, price, image);
@@ -30,7 +30,7 @@ export class MenusService {
         const menus = await this.menusRepository.findMenusByRestaurantId(restaurantId);
 
         if (!menus) {
-            throw { message: "메뉴가 없습니다." };
+            throw new NotFoundError('해당 메뉴가 존재하지 않습니다.');
         }
 
         return menus;
@@ -39,7 +39,7 @@ export class MenusService {
     findOneMenu = async (restaurantId, menuId) => {
         const menu = await this.menusRepository.findMenuByIds(restaurantId, menuId);
         if (!menu) {
-            throw { message: "해당 메뉴가 존재하지 않습니다." };
+            throw new NotFoundError('해당 메뉴가 존재하지 않습니다.');
         }
         return menu;
     };
@@ -54,8 +54,14 @@ export class MenusService {
         }
 
         if (!validation) {
-            throw { code: 400, message: '자신의 식당의 메뉴만 수정할 수 있습니다.' };
+            throw new ValidationError('자신의 식당의 메뉴만 작성할 수 있습니다.');
         }
+
+        const menu = await this.menusRepository.findMenuByIds(restaurantId, menuId);
+        if(!menu){
+            throw new NotFoundError('수정하려는 메뉴가 존재하지 않습니다.');
+        }
+
         await this.menusRepository.updateMenu(restaurantId, menuId, updatedData);
     };
 
@@ -66,6 +72,11 @@ export class MenusService {
 
         if (!comparison) {
             throw new InconsistencyError('비밀번호가 일치하지 않습니다. 다시 입력해주세요.');
+        }
+
+        const menu = await this.menusRepository.findMenuByIds(restaurantId, menuId);
+        if(!menu){
+            throw new NotFoundError('삭제하려는 메뉴가 존재하지 않습니다.');
         }
 
         await this.menusRepository.deleteMenu(restaurantId, menuId);
